Fix the real bug of search function. Do not ignore the FUCKING WARNING of React or else you will lost whole day for some stupid bug!!!
