+++ client/src/api/Job.js
@@ -5,7 +5,6 @@ export function getListAllJob(resolve, reject) {
     let apiUrl = process.env.REACT_APP_API_URL;
 
     axios.get(apiUrl + `/job`).then( (jobs) => {
-        console.log(jobs.data)
         resolve(jobs.data)
     }).catch(reject);
 }
