+++ package.json
@@ -15,6 +15,7 @@
         "concurrently": "^7.4.0",
         "cors": "^2.8.5",
         "datatables.net": "^1.13.1",
+        "eslint": "^5.12.0",
         "express": "^4.18.2",
         "firebase": "^9.15.0",
         "helmet": "^6.0.0",
